src/fraud-detection/requirements.txt
kfp==2.9.0
kfp-kubernetes==1.3.0
kfp-pipeline-spec==0.4.0
kfp-server-api==2.3.0
keras==3.5.0
scikit-learn==1.5.2
numpy==1.26.4
tensorflow==2.17.0
boto3==1.35.24
model-registry==0.2.7a1
kserve==0.14.0rc0